refactor(BA-5528): regroup chat CLI under deployment/chat package
Move chat / chat-config commands and their on-disk storage helpers into
a single sub-package so the layout reflects how they relate, instead of
having loose deployment_chat_*.py modules sitting next to unrelated
top-level v2 helpers.

New layout:

  src/ai/backend/client/cli/v2/deployment/chat/
    __init__.py     # exports chat, chat_config
    commands.py     # both Click commands
    types.py        # DeploymentChatCacheEntry / Cache / Config + errors
    utils.py        # load_*/save_*/mask_token + on-disk constants

Tests follow:

  tests/unit/client/cli/test_deployment_chat_types.py
  tests/unit/client/cli/test_deployment_chat_utils.py

Removed files:

  src/ai/backend/client/cli/v2/deployment/chat.py
  src/ai/backend/client/cli/v2/deployment/chat_config.py
  src/ai/backend/client/cli/v2/deployment_chat_cache.py
  src/ai/backend/client/cli/v2/deployment_chat_config.py
  tests/unit/client/cli/test_deployment_chat_config.py
