+++ src/ai/backend/client/cli/v2/deployment/chat/__init__.py
@@ -1,0 +1,11 @@
+"""``./bai deployment chat`` and ``chat-config`` CLI commands.
+
+Submodules:
+- :mod:`commands` — Click command/group definitions.
+- :mod:`types` — Pydantic models for the on-disk cache and config files.
+- :mod:`utils` — load/save helpers and ``mask_token``.
+"""
+
+from .commands import chat, chat_config
+
+__all__ = ("chat", "chat_config")
